created lfire-app and lofire-net packages. moved common types to lofire-net
