+++ Cargo.toml
@@ -2,5 +2,7 @@
 members = [
   "lofire",
   "lofire-repo",
+  "lofire-net",
   "lofire-node",
+  "lofire-app",
 ]
