+++ lofire-app/Cargo.toml
@@ -1,5 +1,5 @@
 [package]
-name = "lofire-node"
+name = "lofire-app"
 version = "0.1.0"
 edition = "2021"
 
@@ -8,7 +8,6 @@ edition = "2021"
 [dependencies]
 lofire = { path = "../lofire" }
 lofire-net = { path = "../lofire-net" }
-lofire-repo = { path = "../lofire-repo" }
 blake3 = "1.3.1"
 chacha20 = "0.9.0"
 serde = { version = "1.0", features = ["derive"] }
