+++ lofire-app/src/types.rs
@@ -1,0 +1,6 @@
+//! LoFiRe app data types
+//!
+
+use lofire::types::*;
+use lofire_net::types::*;
+use serde::{Deserialize, Serialize};
